face_reco_linux/requirements.txt
Flask==2.3.2
Flask-SocketIO==5.3.4
face_recognition==1.3.0
opencv-python-headless==4.7.0.72
numpy==1.25.0
Pillow==9.5.0
Flask-Cors==5.0.0
redis==4.6.0
requests==2.30.0
